PriorityQueue bug fixed, minor bugs fixed, UI improved

+++ src/main/java/edu/northeastern/course/TheCodeCommandos/Controllers/TableController.java
@@ -4,6 +4,7 @@ import edu.northeastern.course.TheCodeCommandos.Models.Board;
 import edu.northeastern.course.TheCodeCommandos.Models.Card;
 import edu.northeastern.course.TheCodeCommandos.Models.Model;
 import javafx.fxml.Initializable;
+import javafx.geometry.Insets;
 import javafx.scene.control.Button;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.Label;
@@ -36,14 +37,17 @@ public class TableController implements Initializable {
 
     private void setupGridPaneTitle() {
         Label card = new Label("Cards");
-        card.setStyle("-fx-font-size: 18");
+        card.setStyle("-fx-font-size: 16; -fx-font-weight: bold;");
         Label status = new Label("Status");
-        status.setStyle("-fx-font-size: 18");
+        status.setStyle("-fx-font-size: 16; -fx-font-weight: bold;");
         Label dueDate = new Label("Due Date");
-        dueDate.setStyle("-fx-font-size: 18");
+        dueDate.setStyle("-fx-font-size: 16; -fx-font-weight: bold;");
         Label action = new Label("Action");
-        action.setStyle("-fx-font-size: 18");
+        action.setStyle("-fx-font-size: 18; -fx-font-weight: bold;");
         cards_gridpane.addRow(0, card, status, dueDate, action);
+        cards_gridpane.setHgap(110);
+        cards_gridpane.setVgap(50);
+        cards_gridpane.setPadding(new Insets(30, 40, 30, 40));
     }
 
     private void setupGridPane(String boardTitle) {
@@ -54,18 +58,20 @@ public class TableController implements Initializable {
         PriorityQueue<Card> orderedCards = new PriorityQueue<>(Card::compareTo);
         orderedCards.addAll(Model.getInstance().getBoardCards());
 
-        int i = 1;
-        for (Card c: orderedCards) {
-            cards_gridpane.add(new Label(c.getCardName()), 0, i);
-            cards_gridpane.add(new Label(c.getStatus()), 1, i);
-            cards_gridpane.add(new Label(c.dueDateProperty().getValue().toString()), 2, i);
-            Button delete = new Button("Delete");
-            delete.setOnAction(e -> {
-                new Card(c.getCardName(), c.getStatus(), c.dueDateProperty().getValue(), boardTitle).delete();
-                setupGridPane(boardTitle);
-            });
-            cards_gridpane.add(delete, 3, i);
-            i++;
+        int len = orderedCards.size();
+        for (int i = 1; i <= len; i++) {
+            Card c = orderedCards.poll();
+            if (c != null) {
+                cards_gridpane.add(new Label(c.getCardName()), 0, i);
+                cards_gridpane.add(new Label(c.getStatus()), 1, i);
+                cards_gridpane.add(new Label(c.dueDateProperty().getValue().toString()), 2, i);
+                Button delete = new Button("Delete");
+                delete.setOnAction(e -> {
+                    new Card(c.getCardName(), c.getStatus(), c.dueDateProperty().getValue(), boardTitle).delete();
+                    setupGridPane(boardTitle);
+                });
+                cards_gridpane.add(delete, 3, i);
+            }
         }
     }
 }
